+++ fe/src/Pages/Home.tsx
@@ -3,7 +3,7 @@ import styles from "./Home.module.css";
 import { AnimationClass } from "../types/constants";
 import { API_URL } from "../constants";
 import CategoryTab from "../components/CategoryTab/CategoryTab";
-import Main from "../components/Main";
+import Main from "../components/Main/Main";
 
 type HomeProps = {
   changePage: (path: Path) => void;
