+++ fe/src/components/Main/Cart.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useRef, useState } from "react";
-import { API_URL } from "../constants";
-import { PaymentType, Size, Temperature } from "../types/constants";
+import { API_URL } from "../../constants";
+import { PaymentType, Size, Temperature } from "../../types/constants";
 import styles from "./Cart.module.css";
-import Modal from "./Modal";
+import Modal from "./../Modal";
+import CloseButton from "./../CloseButton";
+import { Spinner } from "./../Spinner";
 import { CashPaymentModal, PaymentSelectionModal } from "./Payment";
-import CloseButton from "./CloseButton";
-import { Spinner } from "./Spinner";
 
 type CartProps = {
   cartItems: CartItem[];
